Extract search logic to a separate component
A new Search component is created. It is responsible
for displaying a search input, fetching the results
using a custom fetcher. The search result is displayed
in a custom result item component. The component
also shows a custom loading skeleton and displays
the errors in a toaster. The component also debounces
the search input. The search component is generic,
the response type can be passed as a type argument.

- A new generalized useFetch is created.
- A custom response validator is created for the
spotify API to display only the relevant error message.
- The Shadcn UI toast is added to display errors. Some
customization was needed here to memoize the toast
and dismiss functions to avoid endless loops, when
the toaster is used in a useEffect.
- The search logic is generalized in the AddSong
component. Minor modifications were needed in the
usePlayer hook to handle the different parameters.